按我的理解，应该是这么个结构？
